Adds basic unit tests for happy path and some elements of unhappy path.
This covers the main structural types (repeating & flatten) and the data
type inference for int, float, string, date.  Covers mostly the happy path
and a couple of unhappy paths (eg: what happens when you put a string in
a field expecting an int?).

I probably have the wrong date type here, but come on we all know that
ISO 8601 is the only true way to write dates :-p